+++ packages/react/src/BreakoutBanner/index.ts
@@ -1,1 +1,2 @@
 export * from './BreakoutBanner'
+export * from './BreakoutBanner.types'
